Tidy and move test to more appropriate place.
